Add files via upload
Border_adj.py, takes an image of any aspect ratio and then generates a picture with/without borders for the layout of ModDisplays, depending on the magnetic sensor inputs.

image_generator.py may come in handy to create an image of custom aspect ratios since the 7 inch capacitive touch display comes in odd aspect ratio. I've used it for testing the 2048 by 600 and 2048 by 1200 layouts.