Update border_adj.py
Updated border adjustment file. It outputs two files given an image : the overall image on all the displays as well as the image for the one display we are looking to control for the presentation on Wednesday. Don't forget to change the directories for the three images if you are using this code. 
im - the original image downloaded from the internet
im2_new (inside border_man) - the overall image of all the displays given the magnetic sensor readings
im2_new (main) - the single display we are controlling as of rn. 

Rotation yet to be done